+++ src/main/java/kitchenpos/menu/domain/Menu.java
@@ -1,7 +1,7 @@
 package kitchenpos.menu.domain;
 
-import kitchenpos.menugroup.domain.MenuGroup;
 import kitchenpos.menu.exception.IllegalMenuPriceException;
+import kitchenpos.menugroup.domain.MenuGroup;
 import kitchenpos.menugroup.exception.NoMenuGroupException;
 import kitchenpos.product.domain.Price;
