+++ src/main/java/kitchenpos/order/application/OrderService.java
@@ -1,18 +1,18 @@
 package kitchenpos.order.application;
 
-import kitchenpos.menu.exception.NoMenuException;
-import kitchenpos.table.application.TableService;
 import kitchenpos.menu.application.MenuService;
-import kitchenpos.order.domain.OrderRepository;
 import kitchenpos.menu.domain.Menu;
+import kitchenpos.menu.exception.NoMenuException;
 import kitchenpos.order.domain.Order;
 import kitchenpos.order.domain.OrderLineItem;
-import kitchenpos.table.domain.OrderTable;
+import kitchenpos.order.domain.OrderRepository;
 import kitchenpos.order.dto.OrderLineItemRequest;
 import kitchenpos.order.dto.OrderRequest;
 import kitchenpos.order.dto.OrderResponse;
 import kitchenpos.order.exception.NoOrderException;
 import kitchenpos.order.exception.NoOrderLineItemException;
+import kitchenpos.table.application.TableService;
+import kitchenpos.table.domain.OrderTable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.util.CollectionUtils;
