+++ src/main/java/kitchenpos/order/domain/OrderValidatorImpl.java
@@ -1,16 +1,15 @@
-package kitchenpos.table.application;
+package kitchenpos.order.domain;
 
-import kitchenpos.order.domain.OrderRepository;
-import kitchenpos.order.domain.OrderStatus;
-import kitchenpos.table.domain.OrderTableRepository;
+import kitchenpos.table.domain.OrderValidator;
 import kitchenpos.table.exception.CannotChangeEmptyException;
 import org.springframework.stereotype.Service;
 
 import java.util.Arrays;
 
 @Service
-public class OrderValidator {
+public class OrderValidatorImpl implements OrderValidator {
     OrderRepository orderRepository;
+
     public void canUngroupOrChange(Long id) {
         if (orderRepository.existsByOrderTableIdAndOrderStatusIn(id, Arrays.asList(OrderStatus.COOKING.name(),
                 OrderStatus.MEAL.name()))) {
