+++ src/main/java/kitchenpos/table/domain/OrderValidator.java
@@ -1,0 +1,5 @@
+package kitchenpos.table.domain;
+
+public interface OrderValidator {
+    void canUngroupOrChange(Long id);
+}
