+++ src/main/java/kitchenpos/tablegroup/domain/TableGroupRepository.java
@@ -1,6 +1,5 @@
 package kitchenpos.tablegroup.domain;
 
-import kitchenpos.tablegroup.domain.TableGroup;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
